+++ ITKBrushROIFilter.mm
@@ -140,7 +140,7 @@ ImageType::Pointer CreateImagePointerFromBuffer(unsigned char *buffer, int buffe
 		dstBuf.width = srcbuf.width = bufferWidth;
 		dstBuf.rowBytes = srcbuf.rowBytes = bufferWidth;
 		err = vImageErode_Planar8( &srcbuf, &dstBuf, 0, 0, kernelErode, structuringElementRadius, structuringElementRadius, kvImageDoNotTile); //	
-		if( err) NSLog(@"%d", err);
+		if( err) NSLog(@"%d", (int) err);
 		
 		memcpy(buff, dstBuf.data, bufferWidth*bufferHeight);
 		free( dstBuf.data);
@@ -203,7 +203,7 @@ ImageType::Pointer CreateImagePointerFromBuffer(unsigned char *buffer, int buffe
 		dstBuf.width = srcbuf.width = bufferWidth;
 		dstBuf.rowBytes = srcbuf.rowBytes = bufferWidth;
 		err = vImageDilate_Planar8( &srcbuf, &dstBuf, 0, 0, kernelDilate, structuringElementRadius, structuringElementRadius, kvImageDoNotTile);	//kvImageDoNotTile
-		if( err) NSLog(@"%d", err);
+		if( err) NSLog(@"%d", (int) err);
 		
 		memcpy(buff,dstBuf.data,bufferWidth*bufferHeight);
 		free( dstBuf.data);
